pyproject.toml
[tool.poetry]
name = "deepemotion"
version = "0.1.0"
description = "Deep classification of emotions of fMRI subjects watching Forest Gump"
authors = ["lungerjo <j.lunger@mail.utoronto.ca>"]
readme = "README.md"

packages = [
    { include = "scripts", from = "src" },
    { include = "models", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
datalad = "^1.1.3"
numpy = "^2.1.2"
nibabel = "^5.3.1"
pandas = "^2.2.3"
datalad-installer = "^1.0.5"
torch = "^2.5.0"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
